Refactor SparseML trainer for multiple tasks
